Moved it to the Common namespace + Added the dispatch in Backend
