+++ src/Common/Events/ForkEvents.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Common\Events;
+
+/**
+ * A helper class that contains all possible fork events
+ */
+final class ForkEvents
+{
+    /**
+     * @var string
+     */
+    const FORK_EVENTS_SESSION_ID_CHANGED = 'fork.events.session.id.changed';
+}
